\name{dingfeiPG-package}
\alias{dingfeiPG-package}
\alias{dingfeiPG}
\docType{package}
\title{
\packageTitle{dingfeiPG}
}
\description{
\packageDescription{dingfeiPG}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dingfeiPG}
\packageIndices{dingfeiPG}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{dingfeiPG}

Maintainer: \packageMaintainer{dingfeiPG}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one ~~
~~ per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
